# Conflicts:
#	aro-service/aro-engine/src/main/java/com/altvil/aro/service/network/NetworkDataRequest.java
#	aro-service/aro-engine/src/main/java/com/altvil/aro/service/network/impl/NetworkDataDAO.java
#	aro-service/aro-engine/src/main/java/com/altvil/aro/service/network/impl/NetworkDataServiceImpl.java
#	aro-service/aro-engine/src/main/java/com/altvil/aro/service/optimization/impl/PlanCommandExecutorServiceImpl.java

+++ etl-vz/etl/stage/private/businesses/vz_tam_etl.sh
@@ -9,9 +9,6 @@ TMPDIR=/gisdata/temp/
 UNZIPTOOL=unzip
 export AWS_DEFAULT_REGION=us-east-1
 
-# Cleaning command for new file:
-# cat 20160727_TAM_with_SITEDUNS_WA.txt | tr -d , | tr '\t' ',' | tr '+' '0' | tr '?' '0' > tam_wa.csv
-
 DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd ) # gets directory the script is running from
 
 IFS=',' read -a STATE_ARRAY <<< "${STATE_CODES}"
@@ -23,7 +20,10 @@ cd $GISROOT;
 for STATE in "${STATE_ARRAY[@]}"
 do
 	rm -f ${TMPDIR}/*.*
-	aws s3 cp s3://public.aro/proto/businesses/tam_${STATE}.zip $GISROOT/tam_${STATE}.zip
+	if [ ! -f $GISROOT/tam_${STATE}.zip ]; then
+		aws s3 cp s3://public.aro/proto/businesses/tam_${STATE}.zip $GISROOT/tam_${STATE}.zip
+	fi
+	
 	$UNZIPTOOL tam_${STATE}.zip -d ${TMPDIR}
 	${PSQL} -a -c "SELECT create_vz_tam_table('${STATE}', '${TARGET_SCHEMA_NAME}');"
 	cat /$TMPDIR/tam_${STATE}.csv | ${PSQL} -a -c "COPY ${TARGET_SCHEMA_NAME}.tam_${STATE} FROM STDIN DELIMITER ',' CSV HEADER;"
