+++ etl-vz/etl/stage/reference/households/infousa_households_etl.sh
@@ -10,8 +10,6 @@ export AWS_DEFAULT_REGION=us-east-1
 
 DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd ) # gets directory the script is running from
 
-${PSQL} -a -f $DIR/create_infousa_households.sql
-
 # Use lower case state names. FIPS codes unnecessary here as well.
 IFS=',' read -a STATE_ARRAY <<< "${STATE_CODES}"
 TARGET_SCHEMA_NAME='ref_households'
@@ -21,7 +19,11 @@ cd $GISROOT;
 for STATE in "${STATE_ARRAY[@]}"
 do
 	rm -f ${TMPDIR}/*.*
-	aws s3 cp s3://public.aro/infousa/households_${STATE}.zip $GISROOT/households_${STATE}.zip
+
+	if [ ! -f $GISROOT/$GISROOT/households_${STATE}.zip ]; then
+		aws s3 cp s3://public.aro/infousa/households_${STATE}.zip $GISROOT/households_${STATE}.zip
+	fi
+
 	$UNZIPTOOL households_${STATE}.zip -d ${TMPDIR}
 	${PSQL} -a -c "SELECT create_infousa_households_table('${STATE}', '${TARGET_SCHEMA_NAME}');"
 	${PSQL} -a -c "COPY ${TARGET_SCHEMA_NAME}.infousa_households_${STATE} (address, city, state, zip5, lon, lat, geog) FROM STDIN DELIMITER ',' CSV HEADER;" </$TMPDIR/households_${STATE}.csv
